update redux folder
