+++ src/redux/configStore.js
@@ -1,7 +1,7 @@
 import { combineReducers, createStore, applyMiddleware } from 'redux';
 import logger from 'redux-logger';
 import thunk from 'redux-thunk';
-import covidData from './Data/data';
+import covidData from './Data/reducers';
 
 const rootReducer = combineReducers({
     covidData,
